requirements.txt
##
# Python dependencies
##

ansible>=2.6.1,<2.7
pathlib2>=2.3.0,<3
requests-unixsocket>=0.1.5,<0.2

# "Please pin your dependency to ruamel.yaml<0.15 for production software."
#  
# https://yaml.readthedocs.io/en/latest/basicuse.html
# 
ruamel.yaml<0.15
